feat(api): rename and refactor node definition methods and types for clarity and consistency in handling groups instead of grounds refactor(graph): move useNodesGet hook to a dedicated hooks directory for better organization and maintainability
